Change testing query to timescale time_bucket
Event.select allows modification of the SQL SELECT statement directly
allowing the use of the timescale time_bucket('n {time-period}',
event_time) to be used in the query.
Adding count(*) and project_id to the SELECT gives bucketed counts of
each grouping.
This can then be separated further by a where clause.